+++ spec/yaml_spec.rb
@@ -1,4 +1,0 @@
-describe 'YAML interface' do
-  it 'run tests' do
-  end
-end